server/requirements.txt
Django==1.11.13
djangorestframework==3.8.2
gunicorn==19.8.1
pytz==2018.5
